+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## [Unreleased]
 
+- Add proper support for disabled modules
+- Fix auth required state of status bar
+
 ## [v1.14.0]
 
 - Add the "Ignore this violation" button for violation card of SCA